fix: corregir errores de localización y propiedades de modelos
Cambios principales:
- Renombrar getter 'of' a 'ofLabel' en L10n para evitar conflicto con método estático
- Actualizar método L10n.of() para usar lookupL10n en lugar de Localizations.of
- Corregir referencias a SystemStatus.isRunning → SystemStatus.running
- Corregir referencias a Metrics.avgLatency → Metrics.avgLatencyMs
- Agregar campo 'description' en Strategy.fromJson/toJson
- Corregir tipo de retorno en DoubleExtension.clampValue (num → double)
- Actualizar usos de l10n.of → l10n.ofLabel en backtest y optimization screens
- Agregar documentación completa de localización en LOCALIZATION.md

Archivos afectados:
- lib/l10n/l10n.dart: Renombrar getter y método of()
- lib/l10n/l10n_en.dart: Renombrar 'of' a 'ofLabel'
- lib/l10n/l10n_es.dart: Renombrar 'of' a 'ofLabel'
- lib/screens/dashboard_screen.dart: Corregir propiedades de modelos
- lib/screens/backtest_screen.dart: Actualizar uso de l10n
- lib/screens/optimization_results_screen.dart: Actualizar uso de l10n
- lib/models/strategy.dart: Agregar campo description
- lib/utils/extensions.dart: Corregir tipo de retorno
- LOCALIZATION.md: Guía completa de uso de l10n

+++ lib/utils/extensions.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:intl/intl.dart';
 import '../models/position.dart';
 import 'formatters.dart';
@@ -571,7 +573,7 @@ extension DoubleExtension on double {
   /// (-5.0).clampValue(1, 10); // 1.0
   /// ```
   double clampValue(double min, double max) {
-    return clamp(min, max);
+    return clamp(min, max).toDouble();
   }
 
   /// Verifica si es aproximadamente igual a otro valor
@@ -835,4 +837,4 @@ extension PositionExtension on Position {
 }
 
 // Importación necesaria para pow
-import 'dart:math';
+
